INFRASYS-6459 code review changes. Adds custom user data config. implements check_sit. moves troposphere calidations to template
