+++ configs/custom_user_data.sh
@@ -1,0 +1,3 @@
+#!/bin/bash -xe
+
+# Place your custom configs below
