+++ configs/troposphere.yml
@@ -10,15 +10,17 @@ template_description: 'SIT (Salt Integration Testing) CloudFormation Template'
 ## launch configuration configs
 launch_configuration_name: 'sitLaunchConfiguration'
 instance_type: 'c3.4xlarge'
+# array of security group IDs. Example is shown
 security_groups: 
-#  - 'sg-44dac528'
+#  - 'sg-78shdjs'
 key_name:  
-# AMI ID of amazon-ecs-optimized  
+# AMI ID of amazon-ecs-optimized e.g. 'ami-33679c77'  
 ami_id: 
 ami_url: 'https://aws.amazon.com/marketplace/search/results/ref=dtl_navgno_search_box?page=1&searchTerms=Amazon+ECS-Optimized+Amazon+Linux+AMI'
 
 ## Autoscaling Group configs
 autoscaling_group_name: 'sitAutoScalingGroup'
+# subnet id e.g. 'subnet-6s0e4s02'
 subnet:
 max_size: 20
 min_size: 0
